Unify messaging graph assembly with typed blueprint configuration
