+++ messaging/messaging/src/main/java/io/helidon/messaging/ChannelRegistry.java
@@ -16,55 +16,19 @@
 
 package io.helidon.messaging;
 
-import java.lang.reflect.GenericArrayType;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
-import java.lang.reflect.TypeVariable;
-import java.lang.reflect.WildcardType;
-import java.time.Duration;
-import java.util.ArrayDeque;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Deque;
-import java.util.HashMap;
-import java.util.IdentityHashMap;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
-import java.util.TreeSet;
-import java.util.stream.Collectors;
 
-import io.helidon.common.GenericType;
 import io.helidon.config.Config;
-import io.helidon.config.ConfigSources;
-import io.helidon.config.spi.ConfigNode;
-import io.helidon.faulttolerance.Retry;
-import io.helidon.faulttolerance.RetryConfig;
-import io.helidon.faulttolerance.RetryContext;
-import io.helidon.faulttolerance.RetryException;
-import io.helidon.faulttolerance.RetryOutcome;
-import io.helidon.faulttolerance.SupplierHelper;
-import io.helidon.messaging.spi.IncomingChannel;
-import io.helidon.messaging.spi.MessagingConnector;
-import io.helidon.messaging.spi.OutgoingChannel;
 import io.helidon.service.registry.Service;
 import io.helidon.service.registry.ServiceRegistry;
 
 /**
- * In-memory channel graph assembled from generated consumer registrations.
+ * Service-registry bootstrap for the shared messaging graph.
  */
 @Service.Singleton
 @Service.RunLevel(MessagingRuntime.RUN_LEVEL)
 class ChannelRegistry implements MessagingRuntime {
-    private static final System.Logger LOGGER = System.getLogger(ChannelRegistry.class.getName());
-
-    private final Map<String, MessagingChannel<?>> channels;
-    private final DeliveryEngine deliveryEngine;
     private final DefaultMessagingGraph graph;
 
     @Service.Inject
@@ -73,48 +37,13 @@ class ChannelRegistry implements MessagingRuntime {
                     Config config,
                     ServiceRegistry serviceRegistry,
                     MessagingLifecycleGuard lifecycleGuard) {
-        MessagingConfig messagingConfig = MessagingConfig.builder()
+        this.graph = (DefaultMessagingGraph) MessagingConfig.builder()
                 .serviceRegistry(serviceRegistry)
                 .config(config.get("messaging"))
+                .consumerRegistrations(consumerRegistrations)
+                .emitterRegistrations(emitterRegistrations)
                 .build();
-        this.deliveryEngine = new DeliveryEngine(messagingConfig.execution());
-        this.graph = new DefaultMessagingGraph(deliveryEngine);
         try {
-            validateRegistrationIdentities(consumerRegistrations, emitterRegistrations);
-            validateRegistrationTypeMetadata(consumerRegistrations, emitterRegistrations);
-            Map<String, PayloadContribution> payloadContributions =
-                    validateChannelPayloadContributions(consumerRegistrations, emitterRegistrations);
-            Map<String, List<ConsumerRegistration>> grouped = new HashMap<>();
-            for (ConsumerRegistration registration : consumerRegistrations) {
-                grouped.computeIfAbsent(registration.channel(), _ -> new ArrayList<>())
-                        .add(registration);
-            }
-
-            Map<String, MessagingChannel<?>> channels = new HashMap<>();
-            grouped.forEach((channel, consumers) -> channels.put(channel, createChannel(config, channel, consumers)));
-            new TreeSet<>(messagingConfig.outgoing().keySet())
-                    .forEach(channel -> ensureChannel(config, channels, channel));
-            new TreeSet<>(messagingConfig.incoming().keySet())
-                    .forEach(channel -> ensureChannel(config, channels, channel));
-            this.channels = Map.copyOf(channels);
-
-            Map<String, MessagingConnector> connectors = connectors(messagingConfig.connector());
-            List<OutgoingBinding> outgoingBindings = prepareOutgoingBindings(messagingConfig.outgoing(), connectors);
-            List<IncomingDescriptor> incomingDescriptors = prepareIncomingDescriptors(messagingConfig.incoming(),
-                                                                                     connectors,
-                                                                                     grouped);
-            Set<String> outputChannels = new LinkedHashSet<>(grouped.keySet());
-            outgoingBindings.stream().map(OutgoingBinding::channel).forEach(outputChannels::add);
-            validateGeneratedProducerTargets(consumerRegistrations,
-                                             emitterRegistrations,
-                                             grouped,
-                                             outputChannels);
-            validateFailureRoutes(incomingDescriptors, outputChannels, grouped, payloadContributions);
-            validateIncomingOutputs(incomingDescriptors, outputChannels);
-            registerProcessorRoutes(consumerRegistrations);
-
-            configureOutgoingConnectors(outgoingBindings);
-            configureIncomingConnectors(incomingDescriptors);
             graph.prepare();
         } catch (RuntimeException | Error e) {
             graph.abortPreparation(e);
@@ -123,47 +52,9 @@ class ChannelRegistry implements MessagingRuntime {
         lifecycleGuard.register(this);
     }
 
-    static MessagingExecutionConfig executionConfig(Config root, String channel) {
-        Map<String, String> properties = new LinkedHashMap<>();
-        root.get(MessagingConfigSupport.EXECUTION).detach().asMap().ifPresent(properties::putAll);
-        if (channel != null) {
-            Map<String, String> channelProperties = new LinkedHashMap<>();
-            root.get(MessagingConfigSupport.CHANNEL_PREFIX + Config.Key.escapeName(channel) + ".execution")
-                    .detach()
-                    .asMap()
-                    .ifPresent(channelProperties::putAll);
-            if (channelProperties.containsKey("shutdown-timeout")) {
-                throw new IllegalArgumentException("Channel execution configuration must not override global "
-                                                           + "shutdown-timeout: " + channel);
-            }
-            properties.putAll(channelProperties);
-        }
-        return MessagingExecutionConfig.create(Config.just(ConfigSources.create(properties)));
-    }
-
-    /**
-     * Emit a batch of messages to a named channel.
-     * <p>
-     * This method returns only after every required output completes successfully. Output failures are propagated
-     * unchanged.
-     *
-     * @param channel channel name
-     * @param messages immutable message batch
-     * @param <T> payload type
-     * @throws NullPointerException if {@code channel} or {@code messages} is {@code null}
-     * @throws MessagingException if the channel does not exist
-     * @throws RuntimeException if an output fails
-     */
     @Override
     public <T> void emitBatch(String channel, MessageBatch<? extends T> messages) {
-        Objects.requireNonNull(channel, "channel");
-        Objects.requireNonNull(messages, "messages");
-        MessagingChannel<?> messagingChannel = channels.get(channel);
-        if (messagingChannel == null) {
-            throw new MessagingException("Unknown messaging channel " + channel);
-        }
-        graph.ensureRunning();
-        emitBatch(messagingChannel, messages);
+        graph.emitBatch(channel, messages);
     }
 
     @Service.PreDestroy
@@ -171,14 +62,8 @@ class ChannelRegistry implements MessagingRuntime {
         graph.close();
     }
 
-    /**
-     * Get an assembled channel.
-     *
-     * @param channel channel name
-     * @return channel if it exists
-     */
     Optional<MessagingChannel<?>> channel(String channel) {
-        return Optional.ofNullable(channels.get(channel));
+        return graph.channel(channel);
     }
 
     @Service.PostConstruct
@@ -186,1655 +71,7 @@ class ChannelRegistry implements MessagingRuntime {
         graph.start();
     }
 
-    /**
-     * Add an outgoing connector to a named channel.
-     *
-     * @param channel channel name
-     * @param connector outgoing connector
-     */
-    void addOutgoingConnector(String channel, OutgoingChannel connector) {
-        MessagingChannel<?> messagingChannel = channels.get(channel);
-        if (messagingChannel == null) {
-            throw new IllegalArgumentException("Unknown messaging channel " + channel);
-        }
-        if (!(messagingChannel instanceof DefaultMessagingChannel<?> defaultMessagingChannel)) {
-            throw new IllegalArgumentException("Unsupported channel implementation "
-                                                       + messagingChannel.getClass().getName());
-        }
-        graph.addBinding(connector);
-        defaultMessagingChannel.addOutgoingConnector(connector);
-    }
-
-    /**
-     * Create a context for an incoming connector.
-     *
-     * @param channel channel name
-     * @return incoming connector context
-     */
     IncomingConnectorContext incomingContext(String channel) {
-        return incomingContext(channel, FailurePolicy.create());
+        return new DefaultIncomingConnectorContext(graph, channel, FailurePolicy.create());
     }
-
-    private static String requireConnectorName(Config channelConfig, String direction, String channel) {
-        return channelConfig.get(ConnectorConfigSupport.CONNECTOR_ATTRIBUTE)
-                .asString()
-                .filter(connector -> !connector.isBlank())
-                .orElseThrow(() -> new IllegalArgumentException("Configured " + direction + " channel " + channel
-                                                                        + " must declare a non-blank connector"));
-    }
-
-    private IncomingConnectorContext incomingContext(String channel, FailurePolicy failurePolicy) {
-        return new RegistryIncomingConnectorContext(channel, failurePolicy);
-    }
-
-    private void validateRegistrationIdentities(List<ConsumerRegistration> consumerRegistrations,
-                                                List<EmitterRegistration> emitterRegistrations) {
-        Set<String> handlerIds = new LinkedHashSet<>();
-        for (ConsumerRegistration registration : consumerRegistrations) {
-            String handlerId = requireRegistrationIdentity("Handler", registration.handlerId());
-            if (!handlerIds.add(handlerId)) {
-                throw new IllegalArgumentException("Duplicate messaging handler registration " + handlerId);
-            }
-        }
-
-        Set<String> producerIds = new LinkedHashSet<>();
-        for (EmitterRegistration registration : emitterRegistrations) {
-            String producerId = requireRegistrationIdentity("Producer", registration.producerId());
-            if (!producerIds.add(producerId)) {
-                throw new IllegalArgumentException("Duplicate messaging producer registration " + producerId);
-            }
-        }
-    }
-
-    private String requireRegistrationIdentity(String kind, String identity) {
-        if (identity == null || identity.isBlank()) {
-            throw new IllegalArgumentException(kind + " registration identity must not be blank");
-        }
-        return identity;
-    }
-
-    private void validateRegistrationTypeMetadata(List<ConsumerRegistration> consumerRegistrations,
-                                                  List<EmitterRegistration> emitterRegistrations) {
-        for (ConsumerRegistration registration : consumerRegistrations) {
-            String handler = "Messaging handler " + registration.handlerId();
-            GenericType<?> payloadType = registration.payloadGenericType();
-            GenericType<?> envelopeType = registration.envelopeGenericType();
-            validateType(handler + " payload", payloadType);
-            validateType(handler + " envelope", envelopeType);
-            validateEnvelopePayloadType(handler, payloadType, envelopeType);
-            if (registration instanceof ProcessorRegistration processor) {
-                String outgoing = "Messaging processor " + processor.handlerId() + " outgoing";
-                GenericType<?> outgoingPayloadType = processor.outgoingPayloadGenericType();
-                GenericType<?> outgoingEnvelopeType = processor.outgoingEnvelopeGenericType();
-                validateType(outgoing + " payload", outgoingPayloadType);
-                validateType(outgoing + " envelope", outgoingEnvelopeType);
-                validateEnvelopePayloadType(outgoing, outgoingPayloadType, outgoingEnvelopeType);
-            }
-        }
-        for (EmitterRegistration registration : emitterRegistrations) {
-            String emitter = "Messaging emitter " + registration.producerId();
-            GenericType<?> payloadType = registration.payloadGenericType();
-            GenericType<?> envelopeType = registration.envelopeGenericType();
-            validateType(emitter + " payload", payloadType);
-            validateType(emitter + " envelope", envelopeType);
-            validateEnvelopePayloadType(emitter, payloadType, envelopeType);
-        }
-    }
-
-    private void validateType(String source, GenericType<?> type) {
-        GenericType<?> actualType = Objects.requireNonNull(type, source + " type");
-        Class<?> rawType = actualType.rawType();
-        if (rawType.isPrimitive()) {
-            throw new IllegalArgumentException(source + " raw type must not be primitive: "
-                                                       + rawType.getName());
-        }
-    }
-
-    private void validateEnvelopePayloadType(String source,
-                                             GenericType<?> payloadType,
-                                             GenericType<?> envelopeType) {
-        if (!Message.class.isAssignableFrom(envelopeType.rawType())) {
-            throw new IllegalArgumentException(source + " envelope type " + typeName(envelopeType)
-                                                       + " must implement " + Message.class.getName());
-        }
-        Type resolvedMessageType = resolveSupertype(envelopeType.type(), Message.class);
-        if (!(resolvedMessageType instanceof ParameterizedType messageType)) {
-            return;
-        }
-        Type[] arguments = messageType.getActualTypeArguments();
-        if (arguments.length != 1 || hasUnresolvedType(arguments[0])) {
-            return;
-        }
-        Type envelopePayloadType = arguments[0];
-        if (!equivalentType(payloadType.type(), envelopePayloadType)) {
-            throw new IllegalArgumentException(source + " payload generic type " + typeName(payloadType)
-                                                       + " does not match envelope payload type "
-                                                       + typeName(envelopePayloadType) + " declared by "
-                                                       + typeName(envelopeType));
-        }
-    }
-
-    private boolean hasUnresolvedType(Type type) {
-        if (type instanceof TypeVariable<?>) {
-            return true;
-        }
-        if (type instanceof ParameterizedType parameterizedType) {
-            Type owner = parameterizedType.getOwnerType();
-            if (owner != null && hasUnresolvedType(owner)) {
-                return true;
-            }
-            return Arrays.stream(parameterizedType.getActualTypeArguments()).anyMatch(this::hasUnresolvedType);
-        }
-        if (type instanceof GenericArrayType arrayType) {
-            return hasUnresolvedType(arrayType.getGenericComponentType());
-        }
-        if (type instanceof WildcardType wildcardType) {
-            return Arrays.stream(wildcardType.getLowerBounds()).anyMatch(this::hasUnresolvedType)
-                    || Arrays.stream(wildcardType.getUpperBounds()).anyMatch(this::hasUnresolvedType);
-        }
-        return false;
-    }
-
-    private boolean equivalentType(Type first, Type second) {
-        if (sameType(first, second)) {
-            return true;
-        }
-        Type firstComponent = arrayComponent(first);
-        Type secondComponent = arrayComponent(second);
-        if (firstComponent != null || secondComponent != null) {
-            return firstComponent != null
-                    && secondComponent != null
-                    && equivalentType(firstComponent, secondComponent);
-        }
-        if (first instanceof ParameterizedType firstParameterized
-                && second instanceof ParameterizedType secondParameterized) {
-            if (!sameType(firstParameterized.getRawType(), secondParameterized.getRawType())
-                    || !equivalentNullableType(firstParameterized.getOwnerType(), secondParameterized.getOwnerType())) {
-                return false;
-            }
-            return equivalentTypes(firstParameterized.getActualTypeArguments(),
-                                   secondParameterized.getActualTypeArguments());
-        }
-        if (first instanceof WildcardType firstWildcard && second instanceof WildcardType secondWildcard) {
-            return equivalentTypes(firstWildcard.getLowerBounds(), secondWildcard.getLowerBounds())
-                    && equivalentTypes(firstWildcard.getUpperBounds(), secondWildcard.getUpperBounds());
-        }
-        return false;
-    }
-
-    private boolean equivalentNullableType(Type first, Type second) {
-        return first == null ? second == null : second != null && equivalentType(first, second);
-    }
-
-    private boolean equivalentTypes(Type[] first, Type[] second) {
-        if (first.length != second.length) {
-            return false;
-        }
-        for (int i = 0; i < first.length; i++) {
-            if (!equivalentType(first[i], second[i])) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private Map<String, PayloadContribution> validateChannelPayloadContributions(
-            List<ConsumerRegistration> consumerRegistrations,
-            List<EmitterRegistration> emitterRegistrations) {
-        Map<String, PayloadContribution> payloadTypes = new LinkedHashMap<>();
-        for (ConsumerRegistration registration : consumerRegistrations) {
-            addPayloadContribution(payloadTypes,
-                                   registration.channel(),
-                                   registration.payloadGenericType(),
-                                   "handler " + registration.handlerId());
-            if (registration instanceof ProcessorRegistration processor) {
-                addPayloadContribution(payloadTypes,
-                                       processor.outgoingChannel(),
-                                       processor.outgoingPayloadGenericType(),
-                                       "processor " + processor.handlerId());
-            }
-        }
-        for (EmitterRegistration emitter : emitterRegistrations) {
-            addPayloadContribution(payloadTypes,
-                                   emitter.channel(),
-                                   emitter.payloadGenericType(),
-                                   "emitter " + emitter.producerId());
-        }
-        return Map.copyOf(payloadTypes);
-    }
-
-    private void addPayloadContribution(Map<String, PayloadContribution> payloadTypes,
-                                        String channel,
-                                        GenericType<?> payloadType,
-                                        String source) {
-        if (channel == null || channel.isBlank()) {
-            throw new IllegalArgumentException("Messaging channel contributed by " + source + " must not be blank");
-        }
-        GenericType<?> actualType = Objects.requireNonNull(payloadType, "Payload type contributed by " + source);
-        PayloadContribution contribution = new PayloadContribution(actualType, source);
-        PayloadContribution existing = payloadTypes.putIfAbsent(channel, contribution);
-        if (existing != null && !existing.payloadType().equals(actualType)) {
-            throw new IllegalArgumentException("Channel " + channel + " has conflicting payload types "
-                                                       + typeName(existing.payloadType()) + " from "
-                                                       + existing.source() + " and " + typeName(actualType)
-                                                       + " from " + source);
-        }
-    }
-
-    private String typeName(GenericType<?> type) {
-        return typeName(type.type());
-    }
-
-    private String typeName(Type type) {
-        if (type instanceof Class<?> classType) {
-            return classType.getTypeName();
-        }
-        if (type instanceof GenericArrayType arrayType) {
-            return typeName(arrayType.getGenericComponentType()) + "[]";
-        }
-        if (type instanceof ParameterizedType parameterizedType) {
-            return typeName(parameterizedType.getRawType()) + "<"
-                    + Arrays.stream(parameterizedType.getActualTypeArguments())
-                    .map(this::typeName)
-                    .collect(Collectors.joining(", "))
-                    + ">";
-        }
-        if (type instanceof WildcardType wildcardType) {
-            Type[] lowerBounds = wildcardType.getLowerBounds();
-            if (lowerBounds.length > 0) {
-                return "? super " + Arrays.stream(lowerBounds)
-                        .map(this::typeName)
-                        .collect(Collectors.joining(" & "));
-            }
-            Type[] upperBounds = wildcardType.getUpperBounds();
-            if (upperBounds.length == 0
-                    || (upperBounds.length == 1 && upperBounds[0].equals(Object.class))) {
-                return "?";
-            }
-            return "? extends " + Arrays.stream(upperBounds)
-                    .map(this::typeName)
-                    .collect(Collectors.joining(" & "));
-        }
-        return type.getTypeName();
-    }
-
-    private void validateGeneratedProducerTargets(List<ConsumerRegistration> consumerRegistrations,
-                                                  List<EmitterRegistration> emitterRegistrations,
-                                                  Map<String, List<ConsumerRegistration>> groupedConsumers,
-                                                  Set<String> outputChannels) {
-        for (ConsumerRegistration registration : consumerRegistrations) {
-            if (registration instanceof ProcessorRegistration processor) {
-                validateGeneratedProducerTarget("processor",
-                                                processor.handlerId(),
-                                                processor.outgoingChannel(),
-                                                processor.outgoingPayloadGenericType(),
-                                                processor.outgoingEnvelopeGenericType(),
-                                                groupedConsumers,
-                                                outputChannels);
-            }
-        }
-        for (EmitterRegistration emitter : emitterRegistrations) {
-            validateGeneratedProducerTarget("emitter",
-                                            emitter.producerId(),
-                                            emitter.channel(),
-                                            emitter.payloadGenericType(),
-                                            emitter.envelopeGenericType(),
-                                            groupedConsumers,
-                                            outputChannels);
-        }
-    }
-
-    private void validateGeneratedProducerTarget(String kind,
-                                                 String registrationId,
-                                                 String targetChannel,
-                                                 GenericType<?> payloadType,
-                                                 GenericType<?> envelopeType,
-                                                 Map<String, List<ConsumerRegistration>> groupedConsumers,
-                                                 Set<String> outputChannels) {
-        String target = targetChannel == null ? "" : targetChannel;
-        if (target.isBlank()) {
-            throw new IllegalArgumentException("Messaging " + kind + " " + registrationId
-                                                       + " target channel must not be blank");
-        }
-        if (!channels.containsKey(target)) {
-            throw new IllegalArgumentException("Unknown messaging " + kind + " target channel " + target
-                                                       + " for " + registrationId);
-        }
-        if (!outputChannels.contains(target)) {
-            throw new IllegalArgumentException("Messaging " + kind + " target channel " + target
-                                                       + " has no outputs for " + registrationId);
-        }
-
-        GenericType<?> producedPayload = Objects.requireNonNull(payloadType,
-                                                                 kind + " payload type for " + registrationId);
-        GenericType<?> producedEnvelope = Objects.requireNonNull(envelopeType,
-                                                                  kind + " envelope type for " + registrationId);
-        for (ConsumerRegistration targetConsumer : groupedConsumers.getOrDefault(target, List.of())) {
-            if (!producedPayload.equals(targetConsumer.payloadGenericType())) {
-                throw new IllegalArgumentException("Messaging " + kind + " " + registrationId
-                                                           + " produces payload type " + producedPayload.getTypeName()
-                                                           + " but target channel " + target + " expects "
-                                                           + targetConsumer.payloadGenericType().getTypeName());
-            }
-            if (!typeAccepts(targetConsumer.envelopeGenericType().type(), producedEnvelope.type())) {
-                throw new IllegalArgumentException("Messaging " + kind + " " + registrationId
-                                                           + " produces envelope type " + producedEnvelope.getTypeName()
-                                                           + " that target channel " + target + " cannot accept as "
-                                                           + targetConsumer.envelopeGenericType().getTypeName());
-            }
-        }
-    }
-
-    private void registerProcessorRoutes(List<ConsumerRegistration> consumerRegistrations) {
-        for (ConsumerRegistration registration : consumerRegistrations) {
-            if (registration instanceof ProcessorRegistration processor) {
-                graph.addRoute(processor.channel(), processor.outgoingChannel());
-            }
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private MessagingChannel<?> createChannel(Config config,
-                                              String channel,
-                                              List<ConsumerRegistration> consumerRegistrations) {
-        GenericType<Object> payloadType = (GenericType<Object>) payloadType(channel, consumerRegistrations);
-        DefaultMessagingChannel.Builder<Object> builder = new DefaultMessagingChannel.Builder<>();
-        builder.messagingGraph(graph, channel, executionConfig(config, channel))
-                .payloadType(payloadType);
-        builder.addBatchValidator(messages -> validateMessageTypes(consumerRegistrations, messages));
-        for (ConsumerRegistration consumer : consumerRegistrations) {
-            if (consumer instanceof ProcessorRegistration processor) {
-                builder.addBatchOutput(messages -> processAndRoute(processor, messages));
-            } else {
-                builder.addBatchOutput(consumer::dispatch);
-            }
-        }
-        return builder.build();
-    }
-
-    private void processAndRoute(ProcessorRegistration processor, MessageBatch<?> messages) {
-        MessageBatch<?> result = processor.process(messages);
-        if (result == null) {
-            throw new MessagingException("Messaging processor " + processor.handlerId() + " returned a null batch");
-        }
-        if (!messages.sameDelivery(result)) {
-            throw new MessagingException("Messaging processor " + processor.handlerId()
-                                                 + " did not preserve batch delivery lineage");
-        }
-        Class<?> outgoingEnvelopeType = processor.outgoingEnvelopeGenericType().rawType();
-        Class<?> outgoingPayloadType = processor.outgoingPayloadGenericType().rawType();
-        for (Message<?> message : result) {
-            if (!outgoingEnvelopeType.isInstance(message)) {
-                throw new MessagingException("Messaging processor " + processor.handlerId()
-                                                     + " declared outgoing envelope type "
-                                                     + outgoingEnvelopeType.getName()
-                                                     + " but returned " + message.getClass().getName());
-            }
-            Object entity = message.entity();
-            if (entity == null) {
-                throw new MessagingException("Messaging processor " + processor.handlerId()
-                                                     + " returned a null payload");
-            }
-            if (!outgoingPayloadType.isInstance(entity)) {
-                throw new MessagingException("Messaging processor " + processor.handlerId()
-                                                     + " declared outgoing payload type "
-                                                     + outgoingPayloadType.getName()
-                                                     + " but returned " + entity.getClass().getName());
-            }
-        }
-        MessagingChannel<?> target = channels.get(processor.outgoingChannel());
-        if (target == null) {
-            throw new MessagingException("Unknown messaging processor target channel "
-                                                 + processor.outgoingChannel());
-        }
-        emitBatch(target, result);
-    }
-
-    private MessagingChannel<?> createConfiguredChannel(Config config, String channel) {
-        DefaultMessagingChannel.Builder<Object> builder = new DefaultMessagingChannel.Builder<>();
-        builder.messagingGraph(graph, channel, executionConfig(config, channel))
-                .payloadType(Object.class);
-        return builder.build();
-    }
-
-    private <T> void emitBatch(MessagingChannel<?> messagingChannel,
-                               MessageBatch<? extends T> messages) {
-        if (!(messagingChannel instanceof DefaultMessagingChannel<?> defaultMessagingChannel)) {
-            throw new MessagingException("Unsupported messaging channel implementation "
-                                                 + messagingChannel.getClass().getName());
-        }
-        defaultMessagingChannel.emitBatchObject(messages);
-    }
-
-    private <T> void emitRoutedBatch(MessagingChannel<?> messagingChannel,
-                                     MessageBatch<? extends T> messages) {
-        if (!(messagingChannel instanceof DefaultMessagingChannel<?> defaultMessagingChannel)) {
-            throw new MessagingException("Unsupported messaging channel implementation "
-                                                 + messagingChannel.getClass().getName());
-        }
-        defaultMessagingChannel.emitRoutedBatchObject(messages);
-    }
-
-    private <T> void emitRoutedBatch(String channel, MessageBatch<? extends T> messages) {
-        MessagingChannel<?> messagingChannel = channels.get(channel);
-        if (messagingChannel == null) {
-            throw new MessagingException("Unknown messaging channel " + channel);
-        }
-        emitRoutedBatch(messagingChannel, messages);
-    }
-
-    private GenericType<?> payloadType(String channel, List<ConsumerRegistration> consumerRegistrations) {
-        GenericType<?> payloadType = null;
-        for (ConsumerRegistration consumer : consumerRegistrations) {
-            if (payloadType == null) {
-                payloadType = consumer.payloadGenericType();
-            } else if (!payloadType.equals(consumer.payloadGenericType())) {
-                throw new IllegalArgumentException("Channel " + channel + " has conflicting payload types "
-                                                           + payloadType.getTypeName() + " and "
-                                                           + consumer.payloadGenericType().getTypeName());
-            }
-        }
-        validateEnvelopeTypes(channel, consumerRegistrations);
-        return payloadType == null ? GenericType.OBJECT : payloadType;
-    }
-
-    private void validateEnvelopeTypes(String channel, List<ConsumerRegistration> consumers) {
-        for (int first = 0; first < consumers.size(); first++) {
-            ConsumerRegistration firstConsumer = consumers.get(first);
-            for (int second = first + 1; second < consumers.size(); second++) {
-                ConsumerRegistration secondConsumer = consumers.get(second);
-                if (!compatibleEnvelopeTypes(firstConsumer, secondConsumer)) {
-                    throw new IllegalArgumentException("Channel " + channel + " has conflicting message envelope types "
-                                                               + firstConsumer.envelopeGenericType().getTypeName() + " and "
-                                                               + secondConsumer.envelopeGenericType().getTypeName());
-                }
-            }
-        }
-    }
-
-    private boolean compatibleEnvelopeTypes(ConsumerRegistration first, ConsumerRegistration second) {
-        Type firstType = first.envelopeGenericType().type();
-        Type secondType = second.envelopeGenericType().type();
-        return typeAccepts(firstType, secondType) || typeAccepts(secondType, firstType);
-    }
-
-    private boolean typeAccepts(Type target, Type candidate) {
-        if (sameType(target, candidate)) {
-            return true;
-        }
-        if (target instanceof WildcardType wildcard) {
-            return wildcardAccepts(wildcard, candidate);
-        }
-
-        Class<?> targetRawType = GenericType.create(target).rawType();
-        Class<?> candidateRawType = GenericType.create(candidate).rawType();
-        if (!targetRawType.isAssignableFrom(candidateRawType)) {
-            return false;
-        }
-        if (target instanceof Class<?>) {
-            return true;
-        }
-        if (!(target instanceof ParameterizedType targetParameterized)) {
-            return false;
-        }
-
-        Type resolvedCandidate = resolveSupertype(candidate, targetRawType);
-        if (!(resolvedCandidate instanceof ParameterizedType candidateParameterized)) {
-            return false;
-        }
-        return typeArgumentsAccept(targetParameterized.getActualTypeArguments(),
-                                   candidateParameterized.getActualTypeArguments());
-    }
-
-    private boolean typeArgumentsAccept(Type[] targetArguments, Type[] candidateArguments) {
-        if (targetArguments.length != candidateArguments.length) {
-            return false;
-        }
-        for (int i = 0; i < targetArguments.length; i++) {
-            if (!typeArgumentAccepts(targetArguments[i], candidateArguments[i])) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private boolean typeArgumentAccepts(Type target, Type candidate) {
-        if (sameType(target, candidate)) {
-            return true;
-        }
-        if (target instanceof WildcardType wildcard) {
-            return wildcardAccepts(wildcard, candidate);
-        }
-        if (target instanceof ParameterizedType targetParameterized
-                && candidate instanceof ParameterizedType candidateParameterized
-                && sameType(targetParameterized.getRawType(), candidateParameterized.getRawType())) {
-            return typeArgumentsAccept(targetParameterized.getActualTypeArguments(),
-                                       candidateParameterized.getActualTypeArguments());
-        }
-        Type targetComponent = arrayComponent(target);
-        if (targetComponent != null) {
-            Type candidateComponent = arrayComponent(candidate);
-            return candidateComponent != null && typeArgumentAccepts(targetComponent, candidateComponent);
-        }
-        return false;
-    }
-
-    private Type arrayComponent(Type type) {
-        if (type instanceof GenericArrayType genericArray) {
-            return genericArray.getGenericComponentType();
-        }
-        if (type instanceof Class<?> arrayType && arrayType.isArray()) {
-            return arrayType.getComponentType();
-        }
-        return null;
-    }
-
-    private boolean wildcardAccepts(WildcardType wildcard, Type candidate) {
-        if (candidate instanceof WildcardType) {
-            return false;
-        }
-        for (Type lowerBound : wildcard.getLowerBounds()) {
-            if (!typeAccepts(candidate, lowerBound)) {
-                return false;
-            }
-        }
-        for (Type upperBound : wildcard.getUpperBounds()) {
-            if (!typeAccepts(upperBound, candidate)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private Type resolveSupertype(Type candidate, Class<?> targetRawType) {
-        Class<?> candidateRawType = GenericType.create(candidate).rawType();
-        if (candidateRawType.equals(targetRawType)) {
-            return candidate;
-        }
-
-        Map<TypeVariable<?>, Type> bindings = typeBindings(candidateRawType, candidate);
-        for (Type genericInterface : candidateRawType.getGenericInterfaces()) {
-            Type resolvedInterface = resolveType(genericInterface, bindings);
-            Class<?> interfaceRawType = GenericType.create(resolvedInterface).rawType();
-            if (targetRawType.isAssignableFrom(interfaceRawType)) {
-                Type resolved = resolveSupertype(resolvedInterface, targetRawType);
-                if (resolved != null) {
-                    return resolved;
-                }
-            }
-        }
-        Type genericSuperclass = candidateRawType.getGenericSuperclass();
-        if (genericSuperclass != null) {
-            Type resolvedSuperclass = resolveType(genericSuperclass, bindings);
-            Class<?> superclassRawType = GenericType.create(resolvedSuperclass).rawType();
-            if (targetRawType.isAssignableFrom(superclassRawType)) {
-                return resolveSupertype(resolvedSuperclass, targetRawType);
-            }
-        }
-        return null;
-    }
-
-    private Map<TypeVariable<?>, Type> typeBindings(Class<?> rawType, Type type) {
-        Map<TypeVariable<?>, Type> bindings = new HashMap<>();
-        if (type instanceof ParameterizedType parameterizedType) {
-            TypeVariable<?>[] variables = rawType.getTypeParameters();
-            Type[] arguments = parameterizedType.getActualTypeArguments();
-            for (int i = 0; i < Math.min(variables.length, arguments.length); i++) {
-                bindings.put(variables[i], arguments[i]);
-            }
-        }
-        return bindings;
-    }
-
-    private Type resolveType(Type type, Map<TypeVariable<?>, Type> bindings) {
-        if (type instanceof TypeVariable<?> variable) {
-            return bindings.getOrDefault(variable, variable);
-        }
-        if (type instanceof ParameterizedType parameterizedType) {
-            Type[] arguments = parameterizedType.getActualTypeArguments();
-            Type[] resolvedArguments = new Type[arguments.length];
-            for (int i = 0; i < arguments.length; i++) {
-                resolvedArguments[i] = resolveType(arguments[i], bindings);
-            }
-            Type owner = parameterizedType.getOwnerType();
-            return new ResolvedParameterizedType(owner == null ? null : resolveType(owner, bindings),
-                                                 parameterizedType.getRawType(),
-                                                 resolvedArguments);
-        }
-        if (type instanceof WildcardType wildcardType) {
-            return new ResolvedWildcardType(resolveTypes(wildcardType.getLowerBounds(), bindings),
-                                            resolveTypes(wildcardType.getUpperBounds(), bindings));
-        }
-        if (type instanceof GenericArrayType arrayType) {
-            return new ResolvedGenericArrayType(resolveType(arrayType.getGenericComponentType(), bindings));
-        }
-        return type;
-    }
-
-    private Type[] resolveTypes(Type[] types, Map<TypeVariable<?>, Type> bindings) {
-        Type[] resolved = new Type[types.length];
-        for (int i = 0; i < types.length; i++) {
-            resolved[i] = resolveType(types[i], bindings);
-        }
-        return resolved;
-    }
-
-    private boolean sameType(Type first, Type second) {
-        return first.equals(second) || second.equals(first);
-    }
-
-    private void validateMessageType(ConsumerRegistration consumer, Message<?> message) {
-        Class<?> envelopeType = consumer.envelopeGenericType().rawType();
-        if (!envelopeType.isInstance(message)) {
-            throw new IllegalArgumentException("Channel " + consumer.channel()
-                                                       + " expected message envelope type "
-                                                       + envelopeType.getName()
-                                                       + " but received " + message.getClass().getName());
-        }
-        Object entity;
-        try {
-            entity = message.entity();
-        } catch (MessagingException failure) {
-            if (message instanceof DeadLetterMessage<?>
-                    && envelopeType.isAssignableFrom(DeadLetterMessage.class)) {
-                return;
-            }
-            throw failure;
-        }
-        if (entity == null) {
-            throw new IllegalArgumentException("Channel " + consumer.channel() + " received a null payload");
-        }
-        Class<?> payloadType = consumer.payloadGenericType().rawType();
-        if (!payloadType.isInstance(entity)) {
-            throw new IllegalArgumentException("Channel " + consumer.channel()
-                                                       + " expected payload type "
-                                                       + payloadType.getName()
-                                                       + " but received " + entity.getClass().getName());
-        }
-    }
-
-    private void validateMessageTypes(ConsumerRegistration consumer, MessageBatch<?> messages) {
-        for (Message<?> message : messages.messages()) {
-            validateMessageType(consumer, message);
-        }
-    }
-
-    private void validateMessageTypes(List<ConsumerRegistration> consumers, MessageBatch<?> messages) {
-        for (ConsumerRegistration consumer : consumers) {
-            validateMessageTypes(consumer, messages);
-        }
-    }
-
-    private List<OutgoingBinding> prepareOutgoingBindings(Map<String, Config> configurations,
-                                                          Map<String, MessagingConnector> connectors) {
-        List<OutgoingBinding> bindings = new ArrayList<>();
-        for (String channel : new TreeSet<>(configurations.keySet())) {
-            Config channelConfig = configurations.get(channel);
-            String connectorName = requireConnectorName(channelConfig, "outgoing", channel);
-            MessagingConnector connector = connectors.get(connectorName);
-            if (connector == null) {
-                throw new IllegalArgumentException("No configured connector named " + connectorName
-                                                           + " for outgoing channel " + channel);
-            }
-            bindings.add(new OutgoingBinding(channel, connector, connectorConfig(channelConfig, channel)));
-        }
-        return List.copyOf(bindings);
-    }
-
-    private List<IncomingDescriptor> prepareIncomingDescriptors(
-            Map<String, Config> configurations,
-            Map<String, MessagingConnector> connectors,
-            Map<String, List<ConsumerRegistration>> registrations) {
-        List<IncomingDescriptor> descriptors = new ArrayList<>();
-        for (String channel : new TreeSet<>(configurations.keySet())) {
-            Config channelConfig = configurations.get(channel);
-            String connectorName = requireConnectorName(channelConfig, "incoming", channel);
-            MessagingConnector connector = connectors.get(connectorName);
-            if (connector == null) {
-                throw new IllegalArgumentException("No configured connector named " + connectorName
-                                                           + " for incoming channel " + channel);
-            }
-            FailurePolicy failurePolicy = failurePolicy(channel,
-                                                        channelConfig.get("failure"),
-                                                        registrations.getOrDefault(channel, List.of()));
-            descriptors.add(new IncomingDescriptor(channel,
-                                                   failurePolicy,
-                                                   connector,
-                                                   connectorConfig(channelConfig, channel)));
-        }
-        return List.copyOf(descriptors);
-    }
-
-    private FailurePolicy failurePolicy(String channel,
-                                        Config failureConfig,
-                                        List<ConsumerRegistration> registrations) {
-        List<FailurePolicyContribution> contributions = registrations.stream()
-                .map(registration -> registration.declaredFailurePolicy()
-                        .map(policy -> new FailurePolicyContribution(registration.handlerId(), policy)))
-                .flatMap(Optional::stream)
-                .toList();
-        if (contributions.isEmpty()) {
-            return mergeFailurePolicy(FailurePolicy.create(), failureConfig);
-        }
-
-        FailurePolicyContribution first = contributions.getFirst();
-        FailurePolicy effective = mergeFailurePolicy(first.policy(), failureConfig);
-        for (int i = 1; i < contributions.size(); i++) {
-            FailurePolicyContribution candidate = contributions.get(i);
-            FailurePolicy candidateEffective = mergeFailurePolicy(candidate.policy(), failureConfig);
-            if (!effective.equals(candidateEffective)) {
-                throw new IllegalArgumentException("Incoming channel " + channel
-                                                           + " has conflicting effective failure policies declared by "
-                                                           + "handlers "
-                                                           + first.handlerId() + " and " + candidate.handlerId()
-                                                           + ": " + effective + " and " + candidateEffective);
-            }
-        }
-        return effective;
-    }
-
-    private FailurePolicy mergeFailurePolicy(FailurePolicy declared, Config failureConfig) {
-        FailurePolicy.Builder builder = FailurePolicy.builder(declared);
-        Config retryConfig = failureConfig.get("retry");
-        if (retryConfig.exists()) {
-            builder.retry(RetryConfig.builder(declared.retry()).config(retryConfig).buildPrototype());
-        }
-        FailureDisposition configuredDisposition = failureConfig.get("on-exhausted")
-                .as(FailureDisposition.class)
-                .orElse(null);
-        if (configuredDisposition != null) {
-            builder.onExhausted(configuredDisposition);
-        }
-        Config deadLetterConfig = failureConfig.get("dead-letter");
-        if (deadLetterConfig.exists()) {
-            builder.deadLetter(DeadLetterConfig.create(deadLetterConfig));
-        }
-        if (configuredDisposition != null && configuredDisposition != FailureDisposition.DEAD_LETTER
-                && !failureConfig.get("dead-letter.channel").exists()) {
-            builder.clearDeadLetter();
-        }
-        return builder.build();
-    }
-
-    private void configureOutgoingConnectors(List<OutgoingBinding> bindings) {
-        for (OutgoingBinding binding : bindings) {
-            OutgoingChannel connector = Objects.requireNonNull(
-                    binding.connector().outgoing(binding.config()),
-                    "Outgoing channel result")
-                    .orElseThrow(() -> new IllegalArgumentException("Messaging connector " + binding.connector().name()
-                                                                           + " does not support outgoing channel "
-                                                                           + binding.channel()));
-            addOutgoingConnector(binding.channel(), connector);
-        }
-    }
-
-    private void configureIncomingConnectors(List<IncomingDescriptor> descriptors) {
-        for (IncomingDescriptor descriptor : descriptors) {
-            IncomingConnectorContext context = incomingContext(descriptor.channel(), descriptor.failurePolicy());
-            IncomingChannel connector = Objects.requireNonNull(
-                    descriptor.connector().incoming(descriptor.config()),
-                    "Incoming channel result")
-                    .orElseThrow(() -> new IllegalArgumentException(
-                            "Messaging connector " + descriptor.connector().name()
-                                    + " does not support incoming channel " + descriptor.channel()));
-            graph.addIncomingConnector("connector-" + descriptor.channel(), connector, context);
-        }
-    }
-
-    private void validateIncomingOutputs(List<IncomingDescriptor> bindings, Set<String> outputChannels) {
-        for (IncomingDescriptor binding : bindings) {
-            if (!outputChannels.contains(binding.channel())) {
-                throw new IllegalArgumentException("Incoming channel " + binding.channel() + " has no outputs");
-            }
-        }
-    }
-
-    private void validateFailureRoutes(List<IncomingDescriptor> bindings,
-                                       Set<String> outputChannels,
-                                       Map<String, List<ConsumerRegistration>> consumers,
-                                       Map<String, PayloadContribution> payloadContributions) {
-        Map<String, String> routes = new LinkedHashMap<>();
-        for (IncomingDescriptor binding : bindings) {
-            FailurePolicy policy = binding.failurePolicy();
-            if (policy.onExhausted() != FailureDisposition.DEAD_LETTER) {
-                continue;
-            }
-
-            String source = binding.channel();
-            String target = policy.deadLetter().orElseThrow().channel();
-            if (source.equals(target)) {
-                throw new IllegalArgumentException("Dead-letter channel must not reference itself: " + source);
-            }
-            if (!channels.containsKey(target)) {
-                throw new IllegalArgumentException("Unknown dead-letter channel " + target
-                                                           + " configured for incoming channel " + source);
-            }
-            if (!outputChannels.contains(target)) {
-                throw new IllegalArgumentException("Dead-letter channel " + target
-                                                           + " configured for incoming channel " + source
-                                                           + " has no outputs");
-            }
-            validateDeadLetterPayload(source, target, payloadContributions);
-            validateDeadLetterConsumers(source, target, consumers.getOrDefault(target, List.of()));
-            routes.put(source, target);
-        }
-        validateFailureRouteCycles(routes);
-        routes.forEach(graph::addRoute);
-    }
-
-    private void validateDeadLetterPayload(String source,
-                                           String target,
-                                           Map<String, PayloadContribution> payloadContributions) {
-        PayloadContribution sourcePayload = payloadContributions.get(source);
-        PayloadContribution targetPayload = payloadContributions.get(target);
-        if (sourcePayload != null
-                && targetPayload != null
-                && !sourcePayload.payloadType().equals(targetPayload.payloadType())) {
-            throw new IllegalArgumentException("Dead-letter channel " + target
-                                                       + " has payload type " + typeName(targetPayload.payloadType())
-                                                       + " but incoming channel " + source
-                                                       + " has payload type " + typeName(sourcePayload.payloadType()));
-        }
-    }
-
-    private void validateDeadLetterConsumers(String source,
-                                             String target,
-                                             List<ConsumerRegistration> consumers) {
-        for (ConsumerRegistration consumer : consumers) {
-            Class<?> envelopeType = consumer.envelopeGenericType().rawType();
-            if (!envelopeType.isAssignableFrom(DeadLetterMessage.class)) {
-                throw new IllegalArgumentException("Dead-letter channel " + target
-                                                           + " configured for incoming channel " + source
-                                                           + " has consumer envelope "
-                                                           + consumer.envelopeGenericType().getTypeName()
-                                                           + " that cannot accept "
-                                                           + DeadLetterMessage.class.getName());
-            }
-        }
-    }
-
-    private void validateFailureRouteCycles(Map<String, String> routes) {
-        Set<String> visited = new LinkedHashSet<>();
-        Set<String> visiting = new LinkedHashSet<>();
-        List<String> path = new ArrayList<>();
-        for (String source : routes.keySet()) {
-            visitFailureRoute(source, routes, visited, visiting, path);
-        }
-    }
-
-    private void visitFailureRoute(String source,
-                                   Map<String, String> routes,
-                                   Set<String> visited,
-                                   Set<String> visiting,
-                                   List<String> path) {
-        if (visited.contains(source)) {
-            return;
-        }
-        if (!visiting.add(source)) {
-            int cycleStart = path.indexOf(source);
-            List<String> cycle = new ArrayList<>(path.subList(cycleStart, path.size()));
-            cycle.add(source);
-            throw new IllegalArgumentException("Cyclic dead-letter channel route: " + String.join(" -> ", cycle));
-        }
-        path.add(source);
-
-        String target = routes.get(source);
-        if (target != null) {
-            visitFailureRoute(target, routes, visited, visiting, path);
-        }
-
-        path.removeLast();
-        visiting.remove(source);
-        visited.add(source);
-    }
-
-    private Map<String, MessagingConnector> connectors(List<MessagingConnector> configuredConnectors) {
-        Map<String, MessagingConnector> connectors = new HashMap<>();
-        for (MessagingConnector connector : configuredConnectors) {
-            Objects.requireNonNull(connector, "Messaging connector");
-            String name = Objects.requireNonNull(connector.name(), "Messaging connector name");
-            if (name.isBlank()) {
-                throw new IllegalArgumentException("Messaging connector name must not be blank");
-            }
-            MessagingConnector previous = connectors.putIfAbsent(name, connector);
-            if (previous != null) {
-                throw new IllegalArgumentException("Duplicate messaging connector name " + name);
-            }
-        }
-        return Map.copyOf(connectors);
-    }
-
-    private void ensureChannel(Config config, Map<String, MessagingChannel<?>> channels, String channel) {
-        channels.computeIfAbsent(channel, _ -> createConfiguredChannel(config, channel));
-    }
-
-    private Config connectorConfig(Config channelConfig, String channel) {
-        ConfigNode.ObjectNode source = configObjectNode(channelConfig);
-        ConfigNode.ObjectNode.Builder result = ConfigNode.ObjectNode.builder();
-        source.value().ifPresent(result::value);
-        source.forEach((key, node) -> {
-            if (!key.equals("failure")
-                    && !key.startsWith("failure.")
-                    && !key.equals(ConnectorConfigSupport.CHANNEL_NAME_ATTRIBUTE)
-                    && !key.startsWith(ConnectorConfigSupport.CHANNEL_NAME_ATTRIBUTE + ".")) {
-                result.addNode(key, node);
-            }
-        });
-        return Config.just(ConfigSources.create(result.addValue(ConnectorConfigSupport.CHANNEL_NAME_ATTRIBUTE, channel)
-                                                        .build()));
-    }
-
-    private ConfigNode.ObjectNode configObjectNode(Config config) {
-        if (!config.exists()) {
-            return ConfigNode.ObjectNode.empty();
-        }
-        if (config.type() != Config.Type.OBJECT) {
-            throw new IllegalArgumentException("Channel configuration must be an object: " + config.key());
-        }
-        ConfigNode.ObjectNode.Builder result = ConfigNode.ObjectNode.builder();
-        if (config.hasValue()) {
-            result.value(config.asString().orElseThrow());
-        }
-        config.asNodeList().orElse(List.of()).forEach(child -> result.addNode(
-                Config.Key.escapeName(child.name()),
-                configNode(child)));
-        return result.build();
-    }
-
-    private ConfigNode.ListNode configListNode(Config config) {
-        ConfigNode.ListNode.Builder result = ConfigNode.ListNode.builder();
-        if (config.hasValue()) {
-            result.value(config.asString().orElseThrow());
-        }
-        config.asNodeList().orElse(List.of()).stream()
-                .map(this::configNode)
-                .forEach(node -> {
-                    switch (node.nodeType()) {
-                    case OBJECT -> result.addObject((ConfigNode.ObjectNode) node);
-                    case LIST -> result.addList((ConfigNode.ListNode) node);
-                    case VALUE -> result.addValue((ConfigNode.ValueNode) node);
-                    default -> throw new IllegalStateException("Unsupported configuration node type: " + node.nodeType());
-                    }
-                });
-        return result.build();
-    }
-
-    private ConfigNode configNode(Config config) {
-        return switch (config.type()) {
-        case OBJECT -> configObjectNode(config);
-        case LIST -> configListNode(config);
-        case VALUE -> ConfigNode.ValueNode.create(config.asString().orElseThrow());
-        case MISSING -> throw new IllegalStateException("Cannot copy missing configuration node " + config.key());
-        };
-    }
-
-    private record PendingDelivery(MessageBatch<?> batch,
-                                   boolean initialDelivery,
-                                   int failedAttempts,
-                                   Duration previousDelay,
-                                   long firstCallMillis,
-                                   long startedNanos,
-                                   BatchDeliveryException previousFailure) {
-    }
-
-    private record DeliveryRetryAttempt(int number, Duration previousDelay) {
-    }
-
-    private static final class RetryingDelivery {
-        private MessageBatch<?> batch;
-        private BatchDeliveryException failure;
-
-        private RetryingDelivery(MessageBatch<?> batch) {
-            this.batch = batch;
-        }
-
-        private void failure(MessageBatch<?> batch, BatchDeliveryException failure) {
-            this.batch = batch;
-            this.failure = failure;
-        }
-    }
-
-    private static final class ResolvedParameterizedType implements ParameterizedType {
-        private final Type ownerType;
-        private final Type rawType;
-        private final Type[] actualTypeArguments;
-
-        private ResolvedParameterizedType(Type ownerType, Type rawType, Type[] actualTypeArguments) {
-            this.ownerType = ownerType;
-            this.rawType = rawType;
-            this.actualTypeArguments = actualTypeArguments.clone();
-        }
-
-        @Override
-        public Type[] getActualTypeArguments() {
-            return actualTypeArguments.clone();
-        }
-
-        @Override
-        public Type getRawType() {
-            return rawType;
-        }
-
-        @Override
-        public Type getOwnerType() {
-            return ownerType;
-        }
-
-        @Override
-        public boolean equals(Object object) {
-            if (!(object instanceof ParameterizedType other)) {
-                return false;
-            }
-            return Objects.equals(ownerType, other.getOwnerType())
-                    && rawType.equals(other.getRawType())
-                    && Arrays.equals(actualTypeArguments, other.getActualTypeArguments());
-        }
-
-        @Override
-        public int hashCode() {
-            return Arrays.hashCode(actualTypeArguments)
-                    ^ Objects.hashCode(ownerType)
-                    ^ Objects.hashCode(rawType);
-        }
-    }
-
-    private static final class ResolvedWildcardType implements WildcardType {
-        private final Type[] lowerBounds;
-        private final Type[] upperBounds;
-
-        private ResolvedWildcardType(Type[] lowerBounds, Type[] upperBounds) {
-            this.lowerBounds = lowerBounds.clone();
-            this.upperBounds = upperBounds.clone();
-        }
-
-        @Override
-        public Type[] getUpperBounds() {
-            return upperBounds.clone();
-        }
-
-        @Override
-        public Type[] getLowerBounds() {
-            return lowerBounds.clone();
-        }
-
-        @Override
-        public boolean equals(Object object) {
-            if (!(object instanceof WildcardType other)) {
-                return false;
-            }
-            return Arrays.equals(lowerBounds, other.getLowerBounds())
-                    && Arrays.equals(upperBounds, other.getUpperBounds());
-        }
-
-        @Override
-        public int hashCode() {
-            return Arrays.hashCode(lowerBounds) ^ Arrays.hashCode(upperBounds);
-        }
-    }
-
-    private static final class ResolvedGenericArrayType implements GenericArrayType {
-        private final Type componentType;
-
-        private ResolvedGenericArrayType(Type componentType) {
-            this.componentType = componentType;
-        }
-
-        @Override
-        public Type getGenericComponentType() {
-            return componentType;
-        }
-
-        @Override
-        public boolean equals(Object object) {
-            return object instanceof GenericArrayType other
-                    && componentType.equals(other.getGenericComponentType());
-        }
-
-        @Override
-        public int hashCode() {
-            return componentType.hashCode();
-        }
-    }
-
-    private record PayloadContribution(GenericType<?> payloadType, String source) {
-    }
-
-    private record OutgoingBinding(String channel,
-                                   MessagingConnector connector,
-                                   Config config) {
-    }
-
-    private record FailurePolicyContribution(String handlerId, FailurePolicy policy) {
-    }
-
-    private record IncomingDescriptor(String channel,
-                                      FailurePolicy failurePolicy,
-                                      MessagingConnector connector,
-                                      Config config) {
-    }
-
-    private final class RegistryIncomingConnectorContext implements IncomingConnectorContext {
-        private final String channel;
-        private final FailurePolicy failurePolicy;
-        private final AdmissionTimeoutBudget admissionTimeoutBudget;
-        private final Retry retry;
-
-        private RegistryIncomingConnectorContext(String channel, FailurePolicy failurePolicy) {
-            this.channel = channel;
-            this.failurePolicy = failurePolicy;
-            this.admissionTimeoutBudget = new AdmissionTimeoutBudget(channel, System::nanoTime);
-            this.retry = Retry.create(RetryConfig.builder(failurePolicy.retry())
-                                              .name("messaging-delivery-" + channel)
-                                              .addSkipOn(Error.class)
-                                              .addSkipOn(MessagingRejectedException.class)
-                                              .buildPrototype());
-        }
-
-        @Override
-        public String channel() {
-            return channel;
-        }
-
-        @Override
-        public int maxDeliveryMessages() {
-            return graph.maxDeliveryMessages(channel);
-        }
-
-        @Override
-        public ConnectorDeliveryReservation reserveDelivery() {
-            admissionTimeoutBudget.reset();
-            ConnectorDeliveryReservation reservation = deliveryEngine.reserveConnectorDelivery(
-                    channel,
-                    maxDeliveryMessages(),
-                    this::processDelivery,
-                    this::processFailedDelivery);
-            // Clear a budget started by a concurrent non-blocking attempt before this reservation succeeded.
-            admissionTimeoutBudget.reset();
-            return reservation;
-        }
-
-        @Override
-        public Optional<ConnectorDeliveryReservation> tryReserveDelivery() {
-            return admissionTimeoutBudget.attempt(
-                    () -> deliveryEngine.admissionTimeout(channel)
-                            .map(Duration::toNanos)
-                            .orElse(Long.MAX_VALUE),
-                    remaining -> deliveryEngine.tryReserveConnectorDelivery(
-                            channel,
-                            maxDeliveryMessages(),
-                            remaining,
-                            this::processDelivery,
-                            this::processFailedDelivery));
-        }
-
-        private void processFailedDelivery(MessageBatch<?> root, RuntimeException failure) {
-            Objects.requireNonNull(failure);
-            boolean[] settled = new boolean[root.size()];
-            Deque<PendingDelivery> pending = new ArrayDeque<>();
-            PendingDelivery current = initialDelivery(root);
-            RetryingDelivery retrying = new RetryingDelivery(root);
-            recordDeliveryFailure(root,
-                                  settled,
-                                  pending,
-                                  current,
-                                  retrying,
-                                  new DeliveryRetryAttempt(1, Duration.ZERO),
-                                  failure);
-            handleTerminalFailure(root, settled, retrying.batch, retrying.failure, 1);
-            processPendingDeliveries(root, settled, pending);
-        }
-
-        private void processDelivery(MessageBatch<?> root) {
-            boolean[] settled = new boolean[root.size()];
-            Deque<PendingDelivery> pending = new ArrayDeque<>();
-            pending.addFirst(initialDelivery(root));
-            processPendingDeliveries(root, settled, pending);
-        }
-
-        private PendingDelivery initialDelivery(MessageBatch<?> root) {
-            return new PendingDelivery(root,
-                                       true,
-                                       0,
-                                       Duration.ZERO,
-                                       System.currentTimeMillis(),
-                                       System.nanoTime(),
-                                       null);
-        }
-
-        private void processPendingDeliveries(MessageBatch<?> root,
-                                              boolean[] settled,
-                                              Deque<PendingDelivery> pending) {
-            while (!pending.isEmpty()) {
-                ensureDeliveryActive();
-                PendingDelivery current = pending.removeFirst();
-                processPendingDelivery(root, settled, pending, current);
-            }
-        }
-
-        private void processPendingDelivery(MessageBatch<?> root,
-                                            boolean[] settled,
-                                            Deque<PendingDelivery> pending,
-                                            PendingDelivery current) {
-            Duration remainingTimeout = remainingOverallTimeout(retry.prototype().overallTimeout(),
-                                                                current.startedNanos());
-            if (current.failedAttempts() > 0 && remainingTimeout.isZero()) {
-                handleTerminalFailure(root,
-                                      settled,
-                                      current.batch(),
-                                      Objects.requireNonNull(current.previousFailure()),
-                                      current.failedAttempts());
-                return;
-            }
-            RetryingDelivery retrying = new RetryingDelivery(current.batch());
-            try {
-                retry(current, remainingTimeout).invoke(context -> {
-                    attemptDelivery(root, settled, pending, current, retrying, context);
-                    return Boolean.TRUE;
-                }, this::awaitRetry);
-            } catch (RetryException failure) {
-                handleRetryTermination(root, settled, current, retrying, failure);
-            }
-        }
-
-        private void attemptDelivery(MessageBatch<?> root,
-                                     boolean[] settled,
-                                     Deque<PendingDelivery> pending,
-                                     PendingDelivery current,
-                                     RetryingDelivery retrying,
-                                     RetryContext retryContext) {
-            ensureDeliveryActive();
-            MessageBatch<?> attemptedBatch = retrying.batch;
-            try {
-                ChannelRegistry.this.emitBatch(channel, attemptedBatch);
-                markSettled(root, settled, attemptedBatch);
-                return;
-            } catch (RuntimeException failure) {
-                recordDeliveryFailure(root,
-                                      settled,
-                                      pending,
-                                      current,
-                                      retrying,
-                                      new DeliveryRetryAttempt(retryContext.attempt(), retryContext.previousDelay()),
-                                      failure);
-                throw retryFailure(failure);
-            }
-        }
-
-        private RuntimeException retryFailure(RuntimeException failure) {
-            Throwable current = failure.getCause() instanceof RuntimeException cause ? cause : failure;
-            Set<Throwable> visited = Collections.newSetFromMap(new IdentityHashMap<>());
-            while (current instanceof BatchDeliveryException batchFailure && visited.add(current)) {
-                current = batchFailure.getCause();
-            }
-            if (current instanceof Error error) {
-                throw error;
-            }
-            return SupplierHelper.toRuntimeException(current);
-        }
-
-        private void recordDeliveryFailure(MessageBatch<?> root,
-                                           boolean[] settled,
-                                           Deque<PendingDelivery> pending,
-                                           PendingDelivery current,
-                                           RetryingDelivery retrying,
-                                           DeliveryRetryAttempt retryAttempt,
-                                           RuntimeException failure) {
-            MessageBatch<?> attemptedBatch = retrying.batch;
-            BatchDeliveryException alignedFailure = batchFailure(attemptedBatch, failure);
-            List<Integer> failedIndexes = new ArrayList<>();
-            List<Integer> deferredIndexes = new ArrayList<>();
-            for (int i = 0; i < attemptedBatch.size(); i++) {
-                switch (alignedFailure.outcome(i).status()) {
-                case SUCCEEDED -> markSettled(root, settled, attemptedBatch, i);
-                case NOT_ATTEMPTED -> deferredIndexes.add(i);
-                case FAILED, INDETERMINATE -> failedIndexes.add(i);
-                default -> throw new IllegalStateException("Unsupported batch item status: "
-                                                                   + alignedFailure.outcome(i).status());
-                }
-            }
-
-            if (failedIndexes.isEmpty()) {
-                // A completely deferred dispatch still consumes an attempt so a bounded policy cannot spin forever.
-                failedIndexes.addAll(deferredIndexes);
-                deferredIndexes.clear();
-            } else if (!deferredIndexes.isEmpty()) {
-                MessageBatch<?> deferredBatch = attemptedBatch.subset(deferredIndexes);
-                BatchDeliveryException previousFailure = retrying.failure == null
-                        ? null
-                        : batchFailure(deferredBatch, retrying.failure);
-                pending.addFirst(new PendingDelivery(deferredBatch,
-                                                     false,
-                                                     totalAttempts(current.failedAttempts(),
-                                                                   retryAttempt.number() - 1),
-                                                     retryAttempt.previousDelay(),
-                                                     current.firstCallMillis(),
-                                                     current.startedNanos(),
-                                                     previousFailure));
-            }
-
-            MessageBatch<?> failedBatch = attemptedBatch.subset(failedIndexes);
-            BatchDeliveryException policyFailure = batchFailure(failedBatch, alignedFailure);
-            retrying.failure(failedBatch, policyFailure);
-        }
-
-        private void handleRetryTermination(MessageBatch<?> root,
-                                            boolean[] settled,
-                                            PendingDelivery current,
-                                            RetryingDelivery retrying,
-                                            RetryException retryFailure) {
-            Throwable lastThrowable = retryFailure.outcome().lastThrowable().orElse(null);
-            if (lastThrowable instanceof Error error) {
-                throw error;
-            }
-            MessagingRejectedException rejected = messagingRejection(lastThrowable);
-            if (rejected != null) {
-                throw rejected;
-            }
-
-            RetryOutcome.Termination termination = retryFailure.termination();
-            switch (termination) {
-            case RETRIES_EXHAUSTED, NOT_RETRYABLE, TIMED_OUT -> {
-                BatchDeliveryException failure = Objects.requireNonNull(retrying.failure,
-                                                                        "Missing messaging delivery failure");
-                int attempts = totalAttempts(current.failedAttempts(), retryFailure.outcome().attempts());
-                handleTerminalFailure(root, settled, retrying.batch, failure, attempts);
-            }
-            case INTERRUPTED, CANCELLED -> throw new MessagingRejectedException(
-                    channel,
-                    MessagingRejectedException.Reason.CANCELLED,
-                    "Messaging delivery retry was cancelled on channel " + channel,
-                    retryFailure);
-            case WAIT_FAILED -> throw new MessagingException(
-                    "Messaging delivery retry wait failed on channel " + channel,
-                    retryFailure);
-            case RETRY_POLICY_FAILED -> throw new MessagingException(
-                    "Messaging delivery retry policy failed on channel " + channel,
-                    retryFailure);
-            default -> throw new IllegalStateException("Unsupported retry termination: " + termination);
-            }
-        }
-
-        private void handleTerminalFailure(MessageBatch<?> root,
-                                           boolean[] settled,
-                                           MessageBatch<?> failedBatch,
-                                           BatchDeliveryException policyFailure,
-                                           int failedAttempt) {
-            ensureDeliveryActive();
-            switch (failurePolicy.onExhausted()) {
-            case FAIL -> throw terminalFailure(root,
-                                               settled,
-                                               failedBatch,
-                                               policyFailure,
-                                               "Messaging delivery failed after " + failedAttempt
-                                                       + " attempt(s) on channel " + channel);
-            case DROP -> {
-                ensureDeliveryActive();
-                LOGGER.log(System.Logger.Level.WARNING,
-                           "Messaging delivery failed after " + failedAttempt
-                                   + " attempt(s); dropping " + failedBatch.size()
-                                   + " message(s) from channel " + channel,
-                           policyFailure);
-                markSettled(root, settled, failedBatch);
-            }
-            case DEAD_LETTER -> routeDeadLetters(root,
-                                                settled,
-                                                failedBatch,
-                                                failedAttempt,
-                                                policyFailure);
-            default -> throw new IllegalStateException("Unsupported failure disposition: "
-                                                               + failurePolicy.onExhausted());
-            }
-        }
-
-        private Retry retry(PendingDelivery current, Duration remainingTimeout) {
-            if (current.initialDelivery()) {
-                return retry;
-            }
-            RetryConfig prototype = retry.prototype();
-            if (remainingTimeout.isZero()) {
-                return Retry.create(RetryConfig.builder(prototype)
-                                            .calls(1)
-                                            .overallTimeout(Duration.ofNanos(1))
-                                            .buildPrototype());
-            }
-            Retry.RetryPolicy policy = prototype.retryPolicy().orElseThrow();
-            long previousDelayMillis = durationToMillis(current.previousDelay());
-            Retry.RetryPolicy offsetPolicy = (_, lastDelay, call) -> policy.nextDelayMillis(
-                    current.firstCallMillis(),
-                    call == 1 ? previousDelayMillis : lastDelay,
-                    totalAttempts(current.failedAttempts(), call));
-            return Retry.create(RetryConfig.builder(prototype)
-                                        .retryPolicy(offsetPolicy)
-                                        .overallTimeout(remainingTimeout)
-                                        .buildPrototype());
-        }
-
-        private boolean awaitRetry(Duration delay) {
-            ensureDeliveryActive();
-            try {
-                Thread.sleep(delay);
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new MessagingRejectedException(
-                        channel,
-                        MessagingRejectedException.Reason.CANCELLED,
-                        "Messaging delivery retry was cancelled on channel " + channel,
-                        e);
-            }
-            ensureDeliveryActive();
-            return true;
-        }
-
-        private int totalAttempts(int completedAttempts, int invocationAttempts) {
-            if (completedAttempts > Integer.MAX_VALUE - invocationAttempts) {
-                return Integer.MAX_VALUE;
-            }
-            return completedAttempts + invocationAttempts;
-        }
-
-        private long durationToMillis(Duration duration) {
-            try {
-                return duration.toMillis();
-            } catch (ArithmeticException e) {
-                return Long.MAX_VALUE;
-            }
-        }
-
-        private Duration remainingOverallTimeout(Duration overallTimeout, long startedNanos) {
-            long elapsedNanos = Math.max(0, System.nanoTime() - startedNanos);
-            long overallTimeoutNanos;
-            try {
-                overallTimeoutNanos = overallTimeout.toNanos();
-            } catch (ArithmeticException e) {
-                overallTimeoutNanos = Long.MAX_VALUE;
-            }
-            return Duration.ofNanos(Math.max(0, overallTimeoutNanos - elapsedNanos));
-        }
-
-        private MessagingRejectedException messagingRejection(Throwable throwable) {
-            Throwable current = throwable;
-            for (int i = 0; current != null && i < 64; i++) {
-                if (current instanceof MessagingRejectedException rejected) {
-                    return rejected;
-                }
-                Throwable cause = current.getCause();
-                if (cause == current) {
-                    return null;
-                }
-                current = cause;
-            }
-            return null;
-        }
-
-        private void routeDeadLetters(MessageBatch<?> root,
-                                      boolean[] settled,
-                                      MessageBatch<?> failedBatch,
-                                      int failedAttempt,
-                                      BatchDeliveryException applicationFailure) {
-            String target = failurePolicy.deadLetter().orElseThrow().channel();
-            MessageBatch<?> remaining = failedBatch;
-            while (true) {
-                ensureDeliveryActive();
-                MessageBatch<?> deadLetters = deadLetterBatch(remaining, failedAttempt, applicationFailure);
-                try {
-                    ChannelRegistry.this.emitRoutedBatch(target, deadLetters);
-                    markSettled(root, settled, remaining);
-                    return;
-                } catch (RuntimeException routeFailure) {
-                    ensureDeliveryActive();
-                    BatchDeliveryException alignedRouteFailure = batchFailure(deadLetters, routeFailure);
-                    List<Integer> succeeded = new ArrayList<>();
-                    List<Integer> unresolved = new ArrayList<>();
-                    for (int i = 0; i < remaining.size(); i++) {
-                        if (alignedRouteFailure.outcome(i).status() == BatchItemStatus.SUCCEEDED) {
-                            succeeded.add(i);
-                        } else {
-                            unresolved.add(i);
-                        }
-                    }
-                    if (!succeeded.isEmpty()) {
-                        markSettled(root, settled, remaining.subset(succeeded));
-                    }
-                    if (!succeeded.isEmpty() && !unresolved.isEmpty()) {
-                        remaining = remaining.subset(unresolved);
-                        continue;
-                    }
-
-                    MessageBatch<?> unresolvedBatch = unresolved.isEmpty()
-                            ? remaining
-                            : remaining.subset(unresolved);
-                    BatchDeliveryException unresolvedFailure = unresolved.isEmpty()
-                            ? batchFailure(unresolvedBatch, routeFailure)
-                            : batchFailure(unresolvedBatch, alignedRouteFailure);
-                    BatchDeliveryException result = terminalFailure(
-                            root,
-                            settled,
-                            unresolvedBatch,
-                            unresolvedFailure,
-                            "Dead-letter delivery from channel " + channel + " to channel " + target + " failed");
-                    result.addSuppressed(applicationFailure);
-                    throw result;
-                }
-            }
-        }
-
-        private MessageBatch<?> deadLetterBatch(MessageBatch<?> messages,
-                                                int failedAttempt,
-                                                RuntimeException failure) {
-            List<DeadLetterMessage<Object>> deadLetters = new ArrayList<>(messages.size());
-            for (int i = 0; i < messages.size(); i++) {
-                deadLetters.add(DeadLetterMessage.create(castMessage(messages.get(i)),
-                                                         channel,
-                                                         failedAttempt,
-                                                         deadLetterFailure(messages, i, failure)));
-            }
-            return messages.derive(deadLetters);
-        }
-
-        @SuppressWarnings("unchecked")
-        private Message<Object> castMessage(Message<?> message) {
-            return (Message<Object>) message;
-        }
-
-        private BatchDeliveryException batchFailure(MessageBatch<?> batch, RuntimeException failure) {
-            RuntimeException aligned = BatchDeliveryExceptionSupport.align(batch, failure);
-            return aligned instanceof BatchDeliveryException batchFailure
-                    ? batchFailure
-                    : BatchDeliveryExceptionSupport.indeterminate("Messaging delivery on channel " + channel,
-                                                                  batch,
-                                                                  aligned);
-        }
-
-        private BatchDeliveryException terminalFailure(MessageBatch<?> root,
-                                                       boolean[] settled,
-                                                       MessageBatch<?> unresolved,
-                                                       BatchDeliveryException failure,
-                                                       String message) {
-            List<BatchItemOutcome> outcomes = new ArrayList<>(root.size());
-            for (int i = 0; i < root.size(); i++) {
-                outcomes.add(settled[i]
-                                     ? BatchItemOutcome.succeeded(i)
-                                     : BatchItemOutcome.notAttempted(i));
-            }
-            for (int i = 0; i < unresolved.size(); i++) {
-                int rootIndex = unresolved.lineageIndexIn(root, i);
-                if (rootIndex < 0) {
-                    throw new MessagingException("Delivery failure does not belong to retained batch " + root.id());
-                }
-                BatchItemOutcome outcome = failure.outcome(i);
-                outcomes.set(rootIndex, reindex(rootIndex, outcome));
-            }
-            Throwable cause = failure.getCause() == null ? failure : failure.getCause();
-            BatchDeliveryException result = new BatchDeliveryException(message, cause, root, outcomes);
-            for (Throwable suppressed : failure.getSuppressed()) {
-                result.addSuppressed(suppressed);
-            }
-            return result;
-        }
-
-        private BatchItemOutcome reindex(int index, BatchItemOutcome outcome) {
-            return switch (outcome.status()) {
-            case SUCCEEDED -> BatchItemOutcome.succeeded(index);
-            case FAILED -> outcome.failure()
-                    .map(failure -> BatchItemOutcome.failed(index, failure))
-                    .orElseGet(() -> BatchItemOutcome.failed(index));
-            case NOT_ATTEMPTED -> BatchItemOutcome.notAttempted(index);
-            case INDETERMINATE -> outcome.failure()
-                    .map(failure -> BatchItemOutcome.indeterminate(index, failure))
-                    .orElseGet(() -> BatchItemOutcome.indeterminate(index));
-            default -> throw new IllegalStateException("Unsupported batch item status: " + outcome.status());
-            };
-        }
-
-        private void ensureDeliveryActive() {
-            DeliveryEngine.ensureCurrentDeliveryActive();
-            if (Thread.currentThread().isInterrupted()) {
-                throw new MessagingRejectedException(
-                        channel,
-                        MessagingRejectedException.Reason.CANCELLED,
-                        "Messaging delivery was cancelled on channel " + channel);
-            }
-        }
-
-        private void markSettled(MessageBatch<?> root, boolean[] settled, MessageBatch<?> batch) {
-            if (root.sameDelivery(batch)) {
-                Arrays.fill(settled, true);
-                return;
-            }
-            for (int i = 0; i < batch.size(); i++) {
-                markSettled(root, settled, batch, i);
-            }
-        }
-
-        private void markSettled(MessageBatch<?> root,
-                                 boolean[] settled,
-                                 MessageBatch<?> batch,
-                                 int localIndex) {
-            int rootIndex = batch.lineageIndexIn(root, localIndex);
-            if (rootIndex < 0) {
-                throw new MessagingException("Delivery settlement does not belong to retained batch " + root.id());
-            }
-            settled[rootIndex] = true;
-        }
-
-        private RuntimeException deadLetterFailure(MessageBatch<?> batch,
-                                                   int index,
-                                                   RuntimeException failure) {
-            Set<RuntimeException> visited = Collections.newSetFromMap(new IdentityHashMap<>());
-            RuntimeException current = failure;
-            while (visited.add(current) && current instanceof BatchDeliveryException batchFailure) {
-                int failureIndex = batch.lineageIndexIn(batchFailure.batch(), index);
-                if (failureIndex < 0) {
-                    break;
-                }
-                Throwable itemFailure = batchFailure.outcome(failureIndex).failure().orElse(batchFailure.getCause());
-                if (!(itemFailure instanceof RuntimeException nested)) {
-                    break;
-                }
-                current = nested;
-            }
-            return current;
-        }
-    }
-
 }
