+++ messaging/messaging/src/main/java/io/helidon/messaging/DefaultMessagingGraphBuilder.java
@@ -16,17 +16,9 @@
 
 package io.helidon.messaging;
 
-import java.time.Duration;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.IdentityHashMap;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
-import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Stream;
@@ -37,62 +29,46 @@ import io.helidon.messaging.spi.IncomingChannel;
 import io.helidon.messaging.spi.OutgoingChannel;
 
 /**
- * Default messaging graph builder.
+ * Typed programmatic registrations for the shared messaging configuration.
  */
-final class DefaultMessagingGraphBuilder implements MessagingGraph.Builder {
-    private final List<SourceDefinition> sources = new ArrayList<>();
-    private final Set<Stream<?>> sourceIdentities = Collections.newSetFromMap(new IdentityHashMap<>());
-    private final Set<ChannelConnection> connectorIdentities = Collections.newSetFromMap(new IdentityHashMap<>());
-    private final Map<MessagingChannel<?>, DefaultMessagingChannel<?>> channels = new IdentityHashMap<>();
-    private final Set<MessagingChannel<?>> outputChannels = Collections.newSetFromMap(new IdentityHashMap<>());
-    private final Set<DefaultMessagingChannel<?>> sourceChannels = Collections.newSetFromMap(new IdentityHashMap<>());
-    private final Set<Route> routes = new LinkedHashSet<>();
-    private final Set<Route> directRoutes = new LinkedHashSet<>();
-    private MessagingExecutionConfig executionConfig = MessagingExecutionConfig.builder().build();
-    private DefaultMessagingGraph graph;
+final class DefaultMessagingGraphBuilder extends MessagingGraph.Builder {
     private boolean buildAttempted;
-    private int sourceSequence;
-
-    @Override
-    public MessagingGraph.Builder executionConfig(MessagingExecutionConfig config) {
-        requireUninitialized();
-        executionConfig = Objects.requireNonNull(config);
-        return this;
-    }
 
     @Override
     public <T> MessagingChannel<T> channel(String name, Class<T> payloadType) {
-        return channel(name, GenericType.create(payloadType));
+        return channel(name, GenericType.create(Objects.requireNonNull(payloadType)));
     }
 
     @Override
     public <T> MessagingChannel<T> channel(String name, GenericType<T> payloadType) {
-        return channel(name, payloadType, executionConfig);
+        requireMutable();
+        String channelName = Objects.requireNonNull(name);
+        GenericType<T> actualType = Objects.requireNonNull(payloadType);
+        if (channelName.isBlank()) {
+            throw new IllegalArgumentException("Messaging channel name must not be blank");
+        }
+        if (actualType.rawType().isPrimitive()) {
+            throw new IllegalArgumentException("Messaging channel payload type must not be primitive: "
+                                                       + actualType.getTypeName());
+        }
+        if (channelHandles().stream().anyMatch(channel -> channel.name().equals(channelName))) {
+            throw new IllegalArgumentException("Duplicate messaging channel " + channelName);
+        }
+        MessagingChannel<T> handle = new DefaultMessagingChannelHandle<>(channelName, actualType);
+        addChannelHandle(handle);
+        return handle;
     }
 
     @Override
     public <T> MessagingChannel<T> channel(String name,
                                            GenericType<T> payloadType,
-                                           MessagingExecutionConfig channelExecutionConfig) {
-        requireMutable();
-        String channelName = requireChannelName(name);
-        GenericType<T> channelPayloadType = Objects.requireNonNull(payloadType);
-        if (channelPayloadType.rawType().isPrimitive()) {
-            throw new IllegalArgumentException("Messaging channel payload type must not be primitive: "
-                                                       + channelPayloadType.getTypeName());
-        }
-        MessagingExecutionConfig actualExecutionConfig = Objects.requireNonNull(channelExecutionConfig);
-        if (!executionConfig.shutdownTimeout().equals(actualExecutionConfig.shutdownTimeout())) {
-            throw new IllegalArgumentException("Every channel in a messaging graph must use the same shutdown-timeout");
-        }
-        DefaultMessagingChannel<T> runtimeChannel = new DefaultMessagingChannel.Builder<T>()
-                .payloadType(channelPayloadType)
-                .messagingGraph(graph(), channelName, actualExecutionConfig)
-                .build();
-        MessagingChannel<T> channel = new DefaultMessagingChannelHandle<>(channelName, channelPayloadType);
-        channels.put(channel, runtimeChannel);
-        graph.addEmitter(channel, runtimeChannel);
-        return channel;
+                                           MessagingExecutionConfig executionConfig) {
+        Objects.requireNonNull(executionConfig);
+        MessagingChannel<T> handle = channel(name, payloadType);
+        putChannel(name, MessagingChannelConfig.builder()
+                .execution(executionConfig)
+                .build());
+        return handle;
     }
 
     @Override
@@ -102,51 +78,48 @@ final class DefaultMessagingGraphBuilder implements MessagingGraph.Builder {
 
     @Override
     public <T> MessagingGraph.Builder messageSource(MessagingChannel<T> channel,
-                                                       Stream<? extends Message<? extends T>> source) {
+                                                     Stream<? extends Message<? extends T>> source) {
         return source(channel, source, true);
     }
 
     @Override
     public <T> MessagingGraph.Builder route(MessagingChannel<T> source, MessagingChannel<T> target) {
-        DefaultMessagingChannel<T> actualSource = channel(source);
-        DefaultMessagingChannel<T> actualTarget = channel(target);
+        requireChannel(source);
+        requireChannel(target);
         if (!source.payloadType().equals(target.payloadType())) {
             throw new IllegalArgumentException("Messaging route " + source.name() + " -> " + target.name()
                                                        + " has incompatible payload types "
                                                        + source.payloadType().getTypeName() + " and "
                                                        + target.payloadType().getTypeName());
         }
-        Route route = new Route(source.name(), target.name());
-        if (!directRoutes.add(route)) {
+        if (consumerRegistrations().stream().anyMatch(registration ->
+                registration instanceof RouteRegistration route
+                        && route.channel().equals(source.name())
+                        && route.outgoingChannel().equals(target.name()))) {
             throw new IllegalArgumentException("Duplicate messaging route " + source.name() + " -> " + target.name());
         }
-        routes.add(route);
-        actualSource.addBatchOutput(actualTarget::emitRoutedBatchObject);
-        outputChannels.add(source);
+        addConsumerRegistration(new RouteRegistration(source, target));
         return this;
     }
 
     @Override
     public <I, O> MessagingGraph.Builder payloadProcessor(MessagingChannel<I> source,
-                                                            MessagingChannel<O> target,
-                                                            Function<? super I, ? extends O> processor) {
-        DefaultMessagingChannel<I> actualSource = channel(source);
-        DefaultMessagingChannel<O> actualTarget = channel(target);
-        Function<? super I, ? extends O> actualProcessor = Objects.requireNonNull(processor);
-        actualSource.addBatchOutput(batch -> {
+                                                           MessagingChannel<O> target,
+                                                           Function<? super I, ? extends O> processor) {
+        requireChannel(source);
+        requireChannel(target);
+        Objects.requireNonNull(processor);
+        addConsumerRegistration(new ProcessorDefinition(source, target, batch -> {
             List<Message<O>> results = new ArrayList<>(batch.size());
             for (int i = 0; i < batch.size(); i++) {
                 try {
-                    Message<I> message = batch.get(i);
-                    results.add(Message.create(actualProcessor.apply(source.payloadType().cast(message.entity()))));
+                    results.add(Message.create(processor.apply(source.payloadType().cast(batch.get(i).entity()))));
                 } catch (RuntimeException e) {
                     throw BatchDeliveryExceptionSupport.attemptedPrefix("Messaging payload processor", batch, i, e);
                 }
             }
-            actualTarget.emitBatchObject(batch.derive(results));
-        });
-        routes.add(new Route(source.name(), target.name()));
-        outputChannels.add(source);
+            return batch.derive(results);
+        }));
         return this;
     }
 
@@ -155,121 +128,93 @@ final class DefaultMessagingGraphBuilder implements MessagingGraph.Builder {
             MessagingChannel<I> source,
             MessagingChannel<O> target,
             Function<? super Message<I>, ? extends Message<? extends O>> processor) {
-        DefaultMessagingChannel<I> actualSource = channel(source);
-        DefaultMessagingChannel<O> actualTarget = channel(target);
-        Function<? super Message<I>, ? extends Message<? extends O>> actualProcessor = Objects.requireNonNull(processor);
-        actualSource.addBatchOutput(batch -> {
+        requireChannel(source);
+        requireChannel(target);
+        Objects.requireNonNull(processor);
+        addConsumerRegistration(new ProcessorDefinition(source, target, batch -> {
             List<Message<? extends O>> results = new ArrayList<>(batch.size());
             for (int i = 0; i < batch.size(); i++) {
                 try {
-                    results.add(Objects.requireNonNull(actualProcessor.apply(batch.get(i)), "Message processor result"));
+                    results.add(Objects.requireNonNull(processor.apply(castMessage(batch.get(i))),
+                                                      "Message processor result"));
                 } catch (RuntimeException e) {
                     throw BatchDeliveryExceptionSupport.attemptedPrefix("Messaging message processor", batch, i, e);
                 }
             }
-            actualTarget.emitBatchObject(batch.derive(results));
-        });
-        routes.add(new Route(source.name(), target.name()));
-        outputChannels.add(source);
+            return batch.derive(results);
+        }));
         return this;
     }
 
     @Override
     public <T> MessagingGraph.Builder payloadSink(MessagingChannel<T> source, Consumer<? super T> sink) {
-        DefaultMessagingChannel<T> actualSource = channel(source);
-        Consumer<? super T> actualSink = Objects.requireNonNull(sink);
-        actualSource.addOutput(message -> actualSink.accept(source.payloadType().cast(message.entity())));
-        outputChannels.add(source);
+        requireChannel(source);
+        Objects.requireNonNull(sink);
+        addConsumerRegistration(new ConsumerDefinition(source, batch -> {
+            for (int i = 0; i < batch.size(); i++) {
+                try {
+                    sink.accept(source.payloadType().cast(batch.get(i).entity()));
+                } catch (RuntimeException e) {
+                    throw BatchDeliveryExceptionSupport.sequential("Messaging payload sink", batch, i, e);
+                }
+            }
+        }));
         return this;
     }
 
     @Override
     public <T> MessagingGraph.Builder messageSink(MessagingChannel<T> source,
-                                                     Consumer<? super Message<T>> sink) {
-        DefaultMessagingChannel<T> actualSource = channel(source);
-        Consumer<? super Message<T>> actualSink = Objects.requireNonNull(sink);
-        actualSource.addOutput(message -> actualSink.accept(castMessage(message)));
-        outputChannels.add(source);
+                                                  Consumer<? super Message<T>> sink) {
+        requireChannel(source);
+        Objects.requireNonNull(sink);
+        addConsumerRegistration(new ConsumerDefinition(source, batch -> {
+            for (int i = 0; i < batch.size(); i++) {
+                try {
+                    sink.accept(castMessage(batch.get(i)));
+                } catch (RuntimeException e) {
+                    throw BatchDeliveryExceptionSupport.sequential("Messaging message sink", batch, i, e);
+                }
+            }
+        }));
         return this;
     }
 
     @Override
-    public <T> MessagingGraph.Builder batchSink(MessagingChannel<T> source,
-                                                  Consumer<MessageBatch<T>> sink) {
-        DefaultMessagingChannel<T> actualSource = channel(source);
-        Consumer<MessageBatch<T>> actualSink = Objects.requireNonNull(sink);
-        actualSource.addBatchOutput(actualSink);
-        outputChannels.add(source);
+    public <T> MessagingGraph.Builder batchSink(MessagingChannel<T> source, Consumer<MessageBatch<T>> sink) {
+        requireChannel(source);
+        Objects.requireNonNull(sink);
+        addConsumerRegistration(new ConsumerDefinition(source, batch -> sink.accept(castBatch(batch))));
         return this;
     }
 
     @Override
     public <T> MessagingGraph.Builder incomingChannel(MessagingChannel<T> target, IncomingChannel connection) {
-        DefaultMessagingChannel<T> actualTarget = channel(target);
-        IncomingChannel actualConnection = Objects.requireNonNull(connection);
-        if (sourceChannels.contains(actualTarget)) {
-            throw new IllegalArgumentException("Messaging channel " + target.name() + " already has a source");
-        }
-        if (!connectorIdentities.add(actualConnection)) {
-            throw new IllegalArgumentException("Channel connection is already owned by this messaging graph builder");
-        }
-        try {
-            DefaultMessagingGraph actualGraph = graph();
-            actualGraph.addIncomingConnector(target.name() + "-incoming-" + ++sourceSequence,
-                                             actualConnection,
-                                             new BuilderIncomingConnectorContext(actualGraph, actualTarget));
-        } catch (RuntimeException | Error e) {
-            connectorIdentities.remove(actualConnection);
-            throw e;
-        }
-        sourceChannels.add(actualTarget);
+        requireChannel(target);
+        requireConnection(connection);
+        requireNoSource(target);
+        putIncomingConnection(target, connection);
         return this;
     }
 
     @Override
-    public <T> MessagingGraph.Builder outgoingChannel(MessagingChannel<T> source, OutgoingChannel connector) {
-        DefaultMessagingChannel<T> actualSource = channel(source);
-        OutgoingChannel actualConnector = Objects.requireNonNull(connector);
-        if (!connectorIdentities.add(actualConnector)) {
-            throw new IllegalArgumentException("Outgoing connector is already owned by this messaging graph builder");
-        }
-        try {
-            graph().addBinding(actualConnector);
-        } catch (RuntimeException | Error e) {
-            connectorIdentities.remove(actualConnector);
-            throw e;
-        }
-        actualSource.addOutgoingConnector(actualConnector);
-        outputChannels.add(source);
+    public <T> MessagingGraph.Builder outgoingChannel(MessagingChannel<T> source, OutgoingChannel connection) {
+        requireChannel(source);
+        requireConnection(connection);
+        addConsumerRegistration(new OutgoingDefinition(source, connection));
         return this;
+    }
+
+    @Override
+    public MessagingConfig buildPrototype() {
+        requireMutable();
+        return MessagingConfig.builder().from(this).buildPrototype();
     }
 
     @Override
     public MessagingGraph build() {
-        requireMutable();
+        MessagingConfig config = buildPrototype();
         buildAttempted = true;
-        DefaultMessagingGraph actualGraph = graph();
-        try {
-            validateStreamSourcePaths();
-            validateOutputs();
-            routes.forEach(route -> actualGraph.addRoute(route.source(), route.target()));
-            actualGraph.seal();
-            return actualGraph;
-        } catch (RuntimeException | Error e) {
-            closeAfterBuildFailure(actualGraph, e);
-            throw e;
-        }
-    }
-
-    @Override
-    public void close() {
-        if (buildAttempted) {
-            return;
-        }
-        buildAttempted = true;
-        if (graph != null) {
-            graph.close();
-        }
+        return MessagingGraph.create(config);
     }
 
     @SuppressWarnings("unchecked")
@@ -277,165 +222,197 @@ final class DefaultMessagingGraphBuilder implements MessagingGraph.Builder {
         return (Message<T>) message;
     }
 
-    private <T> MessagingGraph.Builder source(MessagingChannel<T> channel, Stream<?> source, boolean messageSource) {
-        DefaultMessagingChannel<T> actualChannel = channel(channel);
-        Stream<?> actualSource = Objects.requireNonNull(source);
-        if (sourceIdentities.contains(actualSource)) {
-            throw new IllegalArgumentException("Stream source is already owned by this messaging graph builder");
-        }
-        if (sourceChannels.contains(actualChannel)) {
-            throw new IllegalArgumentException("Messaging channel " + channel.name()
-                                                       + " already has a stream source");
-        }
-        String sourceName = channel.name() + "-source-" + ++sourceSequence;
-        Consumer<Object> consumer = messageSource
-                ? value -> actualChannel.emitMessageObject((Message<?>) value)
-                : actualChannel::emitPayloadObject;
-        Runnable streamSource = DefaultMessagingChannel.streamSource(actualSource, consumer);
-        graph().addSource(sourceName, streamSource);
-        sourceIdentities.add(actualSource);
-        sourceChannels.add(actualChannel);
-        sources.add(new SourceDefinition(sourceName, actualChannel));
-        return this;
-    }
-
-    private void validateOutputs() {
-        for (MessagingChannel<?> channel : channels.keySet()) {
-            if (!outputChannels.contains(channel)) {
-                throw new IllegalArgumentException("Messaging channel " + channel.name()
-                                                           + " has no required output");
-            }
-        }
-    }
-
-    private void validateStreamSourcePaths() {
-        Map<String, String> reachedBySource = new LinkedHashMap<>();
-        for (SourceDefinition source : sources) {
-            Set<String> reachable = new LinkedHashSet<>();
-            List<String> pending = new ArrayList<>();
-            pending.add(source.channel().name());
-            while (!pending.isEmpty()) {
-                String channel = pending.removeLast();
-                if (!reachable.add(channel)) {
-                    continue;
-                }
-                for (Route route : routes) {
-                    if (route.source().equals(channel)) {
-                        pending.add(route.target());
-                    }
-                }
-            }
-            for (String channel : reachable) {
-                String previousSource = reachedBySource.putIfAbsent(channel, source.name());
-                if (previousSource != null) {
-                    throw new IllegalArgumentException("Messaging stream source fan-in to channel " + channel
-                                                               + " is not supported; " + previousSource + " and "
-                                                               + source.name() + " converge");
-                }
-            }
-        }
-    }
-
-    private DefaultMessagingGraph graph() {
-        if (graph == null) {
-            graph = new DefaultMessagingGraph(new DeliveryEngine(executionConfig));
-        }
-        return graph;
-    }
-
-    private void requireUninitialized() {
-        requireMutable();
-        if (graph != null) {
-            throw new IllegalStateException("Graph execution must be configured before declaring channels");
-        }
+    @SuppressWarnings("unchecked")
+    private static <T> MessageBatch<T> castBatch(MessageBatch<?> batch) {
+        return (MessageBatch<T>) batch;
     }
 
     private void requireMutable() {
         if (buildAttempted) {
-            throw new IllegalStateException("Messaging graph builder cannot be reused after build or close");
+            throw new IllegalStateException("Messaging graph builder cannot be reused after build");
         }
     }
 
-    private String requireChannelName(String name) {
-        String actualName = Objects.requireNonNull(name);
-        if (actualName.isBlank()) {
-            throw new IllegalArgumentException("Messaging channel name must not be blank");
-        }
-        return actualName;
-    }
-
-    @SuppressWarnings("unchecked")
-    private <T> DefaultMessagingChannel<T> channel(MessagingChannel<T> channel) {
+    private void requireChannel(MessagingChannel<?> channel) {
         requireMutable();
         Objects.requireNonNull(channel);
-        DefaultMessagingChannel<?> defaultChannel = channels.get(channel);
-        if (defaultChannel == null || defaultChannel.graph() != graph()) {
+        if (channelHandles().stream().noneMatch(handle -> handle == channel)) {
             throw new IllegalArgumentException("Messaging channel " + channel.name()
                                                        + " belongs to another messaging graph builder");
         }
-        return (DefaultMessagingChannel<T>) defaultChannel;
     }
 
-    private void closeAfterBuildFailure(DefaultMessagingGraph graph, Throwable failure) {
-        try {
-            graph.close();
-        } catch (RuntimeException | Error closeFailure) {
-            if (failure != closeFailure) {
-                failure.addSuppressed(closeFailure);
-            }
+    private void requireConnection(ChannelConnection connection) {
+        Objects.requireNonNull(connection);
+        if (incomingConnections().values().stream().anyMatch(incoming -> incoming == connection)
+                || consumerRegistrations().stream().anyMatch(registration ->
+                        registration instanceof OutgoingDefinition outgoing && outgoing.connection() == connection)) {
+            throw new IllegalArgumentException("Channel connection is already owned by this messaging graph builder");
         }
     }
 
-    private record SourceDefinition(String name,
-                                    DefaultMessagingChannel<?> channel) {
-    }
-
-    private record Route(String source, String target) {
-    }
-
-    private static final class BuilderIncomingConnectorContext implements IncomingConnectorContext {
-        private final DefaultMessagingGraph graph;
-        private final DefaultMessagingChannel<?> channel;
-        private final AdmissionTimeoutBudget admissionTimeoutBudget;
-
-        private BuilderIncomingConnectorContext(DefaultMessagingGraph graph, DefaultMessagingChannel<?> channel) {
-            this.graph = graph;
-            this.channel = channel;
-            this.admissionTimeoutBudget = new AdmissionTimeoutBudget(channel.name(), System::nanoTime);
+    private void requireNoSource(MessagingChannel<?> channel) {
+        if (emitterRegistrations().stream().anyMatch(registration ->
+                registration instanceof SourceDefinition source && source.handle() == channel)
+                || incomingConnections().containsKey(channel)) {
+            throw new IllegalArgumentException("Messaging channel " + channel.name() + " already has a source");
         }
+    }
 
+    private MessagingGraph.Builder source(MessagingChannel<?> channel, Stream<?> source, boolean messages) {
+        requireChannel(channel);
+        Objects.requireNonNull(source);
+        requireNoSource(channel);
+        if (emitterRegistrations().stream().anyMatch(registration ->
+                registration instanceof SourceDefinition definition && definition.stream() == source)) {
+            throw new IllegalArgumentException("Stream source is already owned by this messaging graph builder");
+        }
+        addEmitterRegistration(new SourceDefinition(channel, source, messages));
+        return this;
+    }
+
+    record SourceDefinition(MessagingChannel<?> handle, Stream<?> stream, boolean messages)
+            implements EmitterRegistration {
         @Override
         public String channel() {
-            return channel.name();
+            return handle.name();
         }
 
         @Override
-        public int maxDeliveryMessages() {
-            return graph.maxDeliveryMessages(channel.name());
+        public String producerId() {
+            return channel() + "-source";
         }
 
         @Override
-        public ConnectorDeliveryReservation reserveDelivery() {
-            admissionTimeoutBudget.reset();
-            ConnectorDeliveryReservation reservation = graph.deliveryEngine().reserveConnectorDelivery(
-                    channel.name(),
-                    maxDeliveryMessages(),
-                    channel::emitBatchObject);
-            admissionTimeoutBudget.reset();
-            return reservation;
+        public GenericType<?> payloadGenericType() {
+            return handle.payloadType();
         }
 
         @Override
-        public Optional<ConnectorDeliveryReservation> tryReserveDelivery() {
-            return admissionTimeoutBudget.attempt(
-                    () -> graph.deliveryEngine().admissionTimeout(channel.name())
-                            .map(Duration::toNanos)
-                            .orElse(Long.MAX_VALUE),
-                    remaining -> graph.deliveryEngine().tryReserveConnectorDelivery(
-                            channel.name(),
-                            maxDeliveryMessages(),
-                            remaining,
-                            channel::emitBatchObject));
+        public GenericType<?> envelopeGenericType() {
+            return GenericType.create(Message.class);
+        }
+    }
+
+    record OutgoingDefinition(MessagingChannel<?> handle, OutgoingChannel connection) implements ConsumerRegistration {
+        @Override
+        public String channel() {
+            return handle.name();
+        }
+
+        @Override
+        public GenericType<?> payloadGenericType() {
+            return handle.payloadType();
+        }
+
+        @Override
+        public GenericType<?> envelopeGenericType() {
+            return GenericType.create(Message.class);
+        }
+
+        @Override
+        public void dispatch(MessageBatch<?> batch) {
+            connection.sendBatch(batch);
+        }
+    }
+
+    private record ConsumerDefinition(MessagingChannel<?> handle, Consumer<MessageBatch<?>> consumer)
+            implements ConsumerRegistration {
+        @Override
+        public String channel() {
+            return handle.name();
+        }
+
+        @Override
+        public GenericType<?> payloadGenericType() {
+            return handle.payloadType();
+        }
+
+        @Override
+        public GenericType<?> envelopeGenericType() {
+            return GenericType.create(Message.class);
+        }
+
+        @Override
+        public void dispatch(MessageBatch<?> batch) {
+            consumer.accept(batch);
+        }
+    }
+
+    private record ProcessorDefinition(MessagingChannel<?> source,
+                                       MessagingChannel<?> target,
+                                       Function<MessageBatch<?>, MessageBatch<?>> processor)
+            implements ProcessorRegistration {
+        @Override
+        public String channel() {
+            return source.name();
+        }
+
+        @Override
+        public GenericType<?> payloadGenericType() {
+            return source.payloadType();
+        }
+
+        @Override
+        public GenericType<?> envelopeGenericType() {
+            return GenericType.create(Message.class);
+        }
+
+        @Override
+        public String outgoingChannel() {
+            return target.name();
+        }
+
+        @Override
+        public GenericType<?> outgoingPayloadGenericType() {
+            return target.payloadType();
+        }
+
+        @Override
+        public GenericType<?> outgoingEnvelopeGenericType() {
+            return GenericType.create(Message.class);
+        }
+
+        @Override
+        public MessageBatch<?> process(MessageBatch<?> batch) {
+            return processor.apply(batch);
+        }
+    }
+
+    record RouteRegistration(MessagingChannel<?> source, MessagingChannel<?> target) implements ProcessorRegistration {
+        @Override
+        public String channel() {
+            return source.name();
+        }
+
+        @Override
+        public GenericType<?> payloadGenericType() {
+            return source.payloadType();
+        }
+
+        @Override
+        public GenericType<?> envelopeGenericType() {
+            return GenericType.create(Message.class);
+        }
+
+        @Override
+        public String outgoingChannel() {
+            return target.name();
+        }
+
+        @Override
+        public GenericType<?> outgoingPayloadGenericType() {
+            return target.payloadType();
+        }
+
+        @Override
+        public GenericType<?> outgoingEnvelopeGenericType() {
+            return GenericType.create(Message.class);
+        }
+
+        @Override
+        public MessageBatch<?> process(MessageBatch<?> batch) {
+            return batch;
         }
     }
 }
